+++ coprocessorInt/src/LMS7002M_lib/LMS7002M_rbb.c
@@ -82,7 +82,7 @@ void LMS7002M_rbb_set_test_out(LMS7002M_t *self, const LMS7002M_chan_t channel, 
     LMS7002M_regs_spi_write(self, 0x0119);
 }
 
-float LMS7002M_rbb_set_pga(LMS7002M_t *self, const LMS7002M_chan_t channel, const float gain)
+double LMS7002M_rbb_set_pga(LMS7002M_t *self, const LMS7002M_chan_t channel, const double gain)
 {
     LMS7002M_set_mac_ch(self, channel);
